pyproject.toml
[project]
name = "Image-background-removal"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Ahmed Mustafa", email = "ahmed.mustafa@imperiumdynamics.com"},
]
dependencies = ["azure-functions>=1.24.0", "pillow>=12.0.0", "numpy>=2.2.6", "opencv-python-headless>=4.12.0.88"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
